fix(windows): use i64 for setLevel and OR collectionBehavior flags
- Change setLevel types from i32 to i64 (NSInteger) in hud_manager, ai/window, notes/window
- Fix collectionBehavior to OR flags instead of overwriting in ai/window and notes/window
- Add FullScreenAuxiliary (256) flag for proper fullscreen space behavior

This ensures proper ABI compatibility on 64-bit macOS and prevents
clobbering flags set by GPUI/AppKit when configuring secondary windows.

+++ src/hud_manager.rs
@@ -620,7 +620,8 @@ fn configure_hud_window_by_bounds(expected_bounds: gpui::Bounds<Pixels>) {
                 // Found it! Configure as HUD overlay
 
                 // Set window level very high (NSPopUpMenuWindowLevel = 101)
-                let hud_level: i32 = 101;
+                // Use i64 (NSInteger) for proper ABI compatibility on 64-bit macOS
+                let hud_level: i64 = 101;
                 let _: () = msg_send![window, setLevel: hud_level];
 
                 // Collection behaviors for HUD:
